requirements.txt
alembic==1.5.2
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
shortuuid==1.0.1
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
WTForms==2.3.3
